Entity-DTO-Mapper

+++ src/main/java/com/theboys/scheduler/dao/AttendanceRepository.java
@@ -3,5 +3,5 @@ package com.theboys.scheduler.dao;
 import com.theboys.scheduler.entity.Attendance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
+public interface AttendanceRepository extends JpaRepository<Attendance, Integer> {
 }
