+++ src/main/java/com/theboys/scheduler/dao/PositionRepository.java
@@ -3,5 +3,5 @@ package com.theboys.scheduler.dao;
 import com.theboys.scheduler.entity.Position;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PositionRepository extends JpaRepository<Position, Long> {
+public interface PositionRepository extends JpaRepository<Position, Integer> {
 }
