+++ src/main/java/com/theboys/scheduler/dao/ScheduleRepository.java
@@ -3,5 +3,5 @@ package com.theboys.scheduler.dao;
 import com.theboys.scheduler.entity.Schedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
+public interface ScheduleRepository extends JpaRepository<Schedule, Integer> {
 }
